requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
datasets>=2.18.0
accelerate>=0.28.0
trl>=0.8.0

# Training
autotrain-advanced>=0.8.0

# Monitoring
wandb>=0.16.0

# Evaluation
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data generation
mistralai>=1.0.0

# HuggingFace Hub
huggingface_hub>=0.21.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
